bot/requirements.txt
# Core Trading Dependencies
binance-connector==3.12.0
supabase==2.3.4
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Additional Tools
requests==2.31.0
schedule==1.2.0
websocket-client==1.6.4
